Install nginx from nginx official or distribution package repos
The distribution package repositories tend to significantly lag
ngnix releases delaying the availability of critical fixes and
new features.  The ngninx project provides package repositories
from which the latest stable and mainline releases can be installed.
This patch adds support for installing from the nginx package
repositories while maintaining the default functionality of installing
from distribution repositories.

The nginx project provides repositories for the following distributions:

  * centos
  * debian
  * rhel
  * ubuntu

To support usage of the correct package repository based on the
distribution, the role was refactored to use distribution specific
variable files and the installation tasks were split out based on
``ansible_os_family``.  The installation of the EPEL repository was also
refactored to use the installation RPMs provided by the Fedora project.
Finally, all bare variables warnings were resolved and
``tasks/main.yml`` was converted to be correct YAML.

**N.B.** ``tasks\ubuntu.yml`` yields a warning regarding the use of
``curl``.  Unfortunately, the version of Python shipped by default with
Ubuntu 14.04 and 16.04 is too old to download the signing key via HTTPS.
Therefore, the ``get_url`` and ``uri`` modules fail with SSL errors.
``curl`` is used as a workaround for this incompatibility.